refactor: :truck: organize interfaces and enums
