+++ src/coordinate/TLE.ts
@@ -2,9 +2,9 @@ import { Earth } from '../body/Earth';
 import { Vector3D } from '../operations/Vector3D';
 import { Sgp4, Sgp4GravConstants } from '../sgp4/sgp4';
 import { EpochUTC } from '../time/EpochUTC';
+import { EciVec3, SatelliteRecord, StateVectorSgp4 } from '../types/types';
 import { DEG2RAD, RAD2DEG, secondsPerDay, TAU } from '../utils/constants';
 import { newtonNu } from '../utils/functions';
-import { EciVec3, SatelliteRecord, StateVectorSgp4 } from './../types/types';
 import { ClassicalElements } from './ClassicalElements';
 import { FormatTle } from './FormatTle';
 import { TEME } from './TEME';
