+++ src/enums/AngularDiameterMethod.ts
@@ -1,0 +1,7 @@
+/**
+ * Enumeration representing different methods for calculating angular diameter.
+ */
+export enum AngularDiameterMethod {
+  Circle,
+  Sphere,
+}
