+++ src/enums/AngularDistanceMethod.ts
@@ -1,0 +1,7 @@
+/**
+ * Enumeration representing different methods for calculating angular distance.
+ */
+export enum AngularDistanceMethod {
+  Cosine,
+  Haversine,
+}
