+++ src/enums/PassType.ts
@@ -1,0 +1,6 @@
+export enum PassType {
+  OUT_OF_VIEW = -1,
+  ENTER = 0,
+  IN_VIEW = 1,
+  EXIT = 2,
+}
