+++ src/enums/index.ts
@@ -1,0 +1,3 @@
+export { AngularDiameterMethod } from './AngularDiameterMethod';
+export { AngularDistanceMethod } from './AngularDistanceMethod';
+export { PassType } from './PassType';
