+++ src/interfaces/OptionsParams.ts
@@ -1,0 +1,3 @@
+export interface OptionsParams {
+  notes: string;
+}
