+++ src/interfaces/index.ts
@@ -1,0 +1,6 @@
+export { BaseObjectParams } from './BaseObjectParams';
+export { GroundPositionParams } from './GroundPositionParams';
+export { OptionsParams } from './OptionsParams';
+export { SatelliteParams } from './SatelliteParams';
+export { SensorParams } from './SensorParams';
+export { StarObjectParams } from './StarObjectParams';
