+++ src/observation/RAE.ts
@@ -1,10 +1,10 @@
 /* eslint-disable no-undefined */
 import { ITRF } from '../coordinate/ITRF';
 import { J2000 } from '../coordinate/J2000';
+import { AngularDistanceMethod } from '../enums/AngularDistanceMethod';
 import { Radians } from '../ootk-core';
 import { Vector3D } from '../operations/Vector3D';
 import { EpochUTC } from '../time/EpochUTC';
-import { AngularDistanceMethod } from '../types/types';
 import { DEG2RAD, halfPi, RAD2DEG, TAU } from '../utils/constants';
 import { angularDistance } from '../utils/functions';
