+++ src/ootk-core.ts
@@ -34,6 +34,8 @@
 export * from './body';
 export * from './coordinate';
 export * from './data/DataHandler';
+export * from './enums';
+export * from './interfaces';
 export * from './objects';
 export * from './observation';
 export * from './operations/operations';
